Fixed include paths - Root level CMakelIst looks at TheodoreDavis/CMakePublic for updated include paths - Moved install location of depedencies to .../cmake/external/ - Added hello_world/include/helloworld path to mimic final install structure -- Needed to avoid build vs install overhead in .cpp file - Refactored hello-world.cpp include paths - Refactored test.cpp include paths
